requirements.txt
flask
flask-sqlalchemy
flask-security-too
flask-restful
flask-cors
flask_caching
argon2_cffi
google-genai
redis